QuNet

QuNet.BasicNode — Type

The default QNode object. Nothing special, but nothing unspecial either ;-)"

QuNet.Coords — Type

Coordinates of the QNode object in two or three spatial dimensions

QuNet.add — Method

Add a QNode object to the network.

Example:
```
using QuNet
Q = QNetwork()
A = BasicNode("A")
add(Q, A)
```

QuNet.percolate_vertices — Method

percolate_vertices(graph, p, exclude)

Perform vertex percolation on graph with probability p, excluding specified vertices.

QuNet.remove — Method

Remove a Qnode object from the network.

The convention for removal echos that of LightGraphs.jl. Suppose a given network has N nodes, and we want to remove the node with the id v.

Check if v < N . If false, simply pop the node from QNetwork.nodes

Else, swap the nodes v and N, then pop from QNetworks.nodes