QuNet

MetaGraphs.MetaDiGraph — Method

Convert a QNetwork object to a directed MetaGraphs in Julia LightGraphs. Directed MetaGraphs are used over weighted di-graphs is because they are more performant at removing edges.

QuNet.BasicChannel — Type

Default Channel type

QuNet.BasicNetwork — Type

Basic QNetwork structure. Although it is mutable, removing nodes and channels should be avoided. Large manipulations are done by parsing QuNet into more optimised graph structures via convertNet!()

QuNet.BasicNode — Type

The default QNode object. Nothing special, but nothing unspecial either ;-)"

QuNet.CartCoords — Type

Coordinates of the QNode object in up to three spatial dimensions

QuNet.CartNode — Type

Similar to BasicNode, but with an extra parameter for cartesian spatial coordinates

QuNet.CartSatNode — Type

Satellite Node in Cartesian Coordinates

QuNet.CartVelocity — Type

Cartesian Velocity in up to 3 spatial coordinates

QuNet.Costs — Type

Cost object for QNodes and QChannels

QuNet.FibreChannel — Type

Fibre optic channel, where the cost is determined by exponential loss in length

QuNet.Pathset — Type

Pathset is a collection of paths that all start at some vertex u and end at some vertex v. The purpose of the pathset is to organise / keep track of paths that are purified together.

QuNet.PurifiedChannel — Type

Channel class for multiple channels purified together

QuNet.E_to_dE — Method

Convert from metric form to decibelic loss

QuNet.F_to_dF — Method

Convert from bell pair fidelity to decibelic form

QuNet.GridNetwork — Method

GridNetwork(dim::Int64, dimY::Int64)

Generates an X by Y grid network.

QuNet.addChannel! — Function

Add one or more channels to the network.

TODO: Throw an error if src or dst not in the network

QuNet.addCostPrefix — Method

Add prefix "CostΔ" onto a cost name

QuNet.addNode! — Method

Add one or more nodes to the network.

If the node being initialised already exists in the network, (the IDs are equal) addNode replaces it. If node ID is greater than the number of nodes in the network, function warns user and sets node ID to nv + 1

QuNet.addPurifyChannel — Method

After purification, this function adds a channel between the src and dst with the purified costs and information about the Pathset.

QuNet.add_async_nodes! — Method

Add asynchronus nodes to a Temporal Network.

This function adds tempnet.nv nodes to the graph. If a given node is source, outgoing asynchronus edges are added to the node that connects it to its temporal counterparts. This is done with incremental timeweights ϵ so that earlier times are prioritized.

Likewise if a node is a dest, incoming asynchronus edges are added with incremental weights ϵ.

QuNet.airCosts — Method

Calculate costs for AirChannel given length

QuNet.ave_paths_used — Method

Given a tally for the number of paths used by each end-user in a greedy_protocol: (i.e. [3,4,2,1] meaning 3 end-users used no paths, 4, end-users used 1 path, etc.) This function finds the average number of paths used in the protocol.

QuNet.c_addQChannel — Method

Add a QChannel to the Network graph and return the vertex corresponding to the channel

QuNet.c_addQNode! — Function

Add a QNode to the Network graph

QuNet.cartDistance — Method

Cartesian distance between two nodes

QuNet.dE_to_E — Method

Convert from decibelic loss to metric form

QuNet.dF_to_F — Method

Convert from decibelic bell pair fidelity to metric form

QuNet.dict_average — Method

Return average average value for an array of cost vectors. If the length of the input is less than 1, mean is not well defined, the key values returned are 'nothing'.

QuNet.dict_err — Method

Return average standard error for an array of cost vectors. If the length of the input is less than 2, error is not well defined, the key values returned are 'nothing'.

QuNet.fibreCosts — Function

Calculate costs for FibreChannel given length Length and attenuation parameter β have unspecified units.

QuNet.findPathInPathset — Method

If a path is in the pathset, return its index in the list. Else return 0.

QuNet.fix_async_nodes_in_time! — Method

For a given node at time T in a TemporalGraph, this function fixes its asynchronus counterpart with respect to T This is done by removing all edges of the asynchronus node except for the one connecting at time T.

QuNet.g_addEdge! — Method

Add an edge to the MetaDiGraph and instantiate the default properties from the edgeProps dictionary.

QuNet.g_addVertex! — Method

Add a vertex to the MetaDiGraph and instantiate the default properties from the vertexProps dictionary.

QuNet.g_edgeCosts — Method

Get the costs associated with an edge in the MetaDiGraph

QuNet.g_getNode — Method

Given a vertex in the graph, get the QNode corresponding to it. If the vertex does not correspond to a QNode, return 0. Else if the vertex doesn't exist in the network, return an error.

QuNet.g_getProp — Method

Wrapper function for getting properties from meta graph.

QuNet.g_getVertex — Method

Given a QNode's id, get the vertex corresponding to it. If no vertex exists, return 0.

QuNet.g_pathCosts — Method

Given a path in the MetaDiGraph, return the associated costs

QuNet.g_removeInactiveChannels! — Method

Remove the inactive channels of a graph

QuNet.g_removeInactiveNodes — Method

Remove the inactive Nodes of a graph

QuNet.g_shortestPath — Method

Find the shortest path in terms of the specified cost field. Returns the network path along with the associated costs.

QuNet.getChannel — Method

Get the channel between src and dst. If no channel exists, return nothing

QuNet.getChannelIdx — Method

Fetch the net.channel index of the channel between src and dst If no channel exists, returns nothing

QuNet.getNode — Method

Get the node corresponding to the given id. If no node exists, returns nothing

QuNet.greedyMultiPath! — Method

greedyMultiPath! is an entanglement routing strategy for multiple user-pairs in a quantum network.

Each user pair has an associated Pathset that starts out empty. For each

userpair, the shortest path between them is found in terms of dF and the path is added to the pathset. If no path exists, nothing is added. This continues until one of the following criteria is met: + The pathsets are filled with the threshold number of paths (given by maxpaths)

Entanglement purification is performed over all the pathsets, removing the

paths without adding purification channels.

If a pathset is empty, no path was found between the end-user pair.

When all the bundles are filled up to the maximum number "maxpaths", use an

entanglement purification method between the path costs. Add the resulting cost vector to the array purpaths. If no paths exist for a given bundle, increment the collision count by one, and add "nothing" to purpaths

Return the array of purified cost vectors and the collision count.

QuNet.greedy_multi_path! — Function

greedymultipath! is an entanglement routing strategy for a quantum network with n end user pairs.

Each user pair has an associated "bundle" of paths that starts out empty.

For each user pair, find the shortest path between them, and add its cost to the bundle. If no path exists, add "nothing" to the bundle.

When all the bundles are filled up to the maximum number "maxpaths", use an

entanglement purification method between the path costs. Add the resulting cost vector to the array purpaths. If no paths exist for a given bundle, increment the collision count by one, and add "nothing" to purpaths

Return the array of purified cost vectors and the collision count.

QuNet.handleRepeatPurification — Method

Checks to see if an identical purification has already been done before so the purification channel isn't duplicated

QuNet.heat_data — Method

Takes a network as input and returns greedymultipath! heatmap data for some number of user pairs. (i.e. a list of efficiency fidelity coordinates for all end users over all num_trials)

QuNet.make_user_pairs — Method

Generate a list of end-users for a TemporalGraph. srclayer and dstlayer specify the temporal locations of the source and dst nodes respectively. The default value for these is -1, which indicates the end-users should be asynchronus.

QuNet.make_user_pairs_wbuffer — Method

Generate a list of user_pairs for a square grid network, ensuring that there is an edge buffer of a given thickness.

QuNet.mapNodeToVert! — Method

Map a node in the QuNet to a vertex in the graph

QuNet.max_timedepth — Method

Find the maximum timedepth reached by a given pathset

QuNet.n_channelCosts — Method

Return the costs for a channel

QuNet.n_getChannels — Method

Get a list of channel vertices between two nodes of the network. Return an empty list if no path is found.

QuNet.n_hasChannel — Function

Given the src and dst of a QChannel, determine if it exists in the MetaDiGraph

QuNet.n_hasNode — Method

Given a QNode, determine whether or not it exists in the MetaDiGraph

QuNet.n_pathCapacity — Method

Return the maximum capacity allowed by a given path

QuNet.n_remAllChannels! — Function

Reduce the channel capacity associated with all channels connecting src and dst by one. If the channel capacity is zero, remove the channel.

QuNet.n_remChannel! — Method

Reduce the channel capacity associated with an channel by one. If the channel capacity is zero, remove the corresponding edge. If the channel is undirected, operate on both directions.

QuNet.n_remNode — Method

Remove a QNode from the graph and reindex the :nodeToVert map accordingly

QuNet.n_remShortestPath! — Method

Remove the shortest path in terms of the specified cost field. Returns the network path along with the associated costs.

QuNet.n_removeVertexPath! — Method

Given a literal path in the MetaGraph, remove it from the network, taking care that channel bandwidth is reduced where applicable and making sure not to remove edges not corresponding to channels.

QuNet.n_setChannelCosts — Method

Set the costs for a channel

QuNet.n_shortestPath — Method

Find the shortest path in terms of the specified cost field. Returns the network path along with the associated costs.

QuNet.n_uniqueChannel — Method

Fetches the channel between srcNode and dstNode provided that one unique channel exists between them. If not, this function throws an error

QuNet.n_uniqueVertexPath — Method

Returns a list of tuples representing a vertex path in the network provided that each channel is unique. If not, this function throws an error.

QuNet.n_vertexToNetPath — Method

Given a path in the MetaDiGraph, convert it to a path in the QNetwork

QuNet.net_performance — Method

Takes a network as input and return greedymultipath! performance statistics for some number of random user pairs. Ensure graph is refreshed before starting.

QuNet.node_timedepth — Method

Get the temporal depth of a node from its index. If asynchronus, return -1 Throws an exception if the node is out of bounds.

QuNet.pathCosts — Method

Get the costs of a path in the network.

QuNet.percolate_vertices — Method

percolate_vertices(graph, p, exclude)

Perform vertex percolation on graph with probability p, excluding specified vertices.

QuNet.plotNetworkGraph — Method

Plot the entire graph structure representing the quantum network

QuNet.purify — Method

Reduce the graph associated with the QuNetwork by performing a purification over a pathset

QuNet.purifyCosts — Method

Calculate the purification costs given a Pathset

QuNet.relative_index — Method

Get the relative index of a node from its original index. ie. for a graph with nv = 10, steps = 3, then node 14 is the same relative to node 4. Another way to think about this is that it gets the index of the corresponding node in the top layer.

QuNet.remChanVert — Method

Remove a vertex corresponding to a channel in the graph.

QuNet.remCostPrefix — Method

Strip prefix "CostΔ" from cost name

QuNet.remPathset! — Method

Remove the pathset from the graph.

If the frequency of a path in the pathset exceeds the capacity of the path, this function sets the frequency to this maximum capacity and removes that amount of paths instead.

QuNet.remove_async_nodes! — Method

Remove all asynchronus nodes of a TemporalGraph.

i.e. those put in place by addasyncnodes!

QuNet.reversePathset — Method

Make a reversed instance of a given Pathset

QuNet.setChanAttrs — Method

Set the attributes of a vertex corresponding to a qchannel. (Essentially all of the attributes of the qchannel except for the costs which go to the edges)

QuNet.setEdgeCosts — Method

Set the costs of a directed edge corresponding to a qchannel. (Because there are two of these in every channel, the costs are set to half their usual value)

QuNet.setVertexAttrs — Method

Set the attrtibutes of a vertex given a qnode and its index

QuNet.temp_to_async — Method

Convert temporal nodes into asynchronus nodes If the node is already asynchronus, it is returned Throws an exception if the node is out of bounds

QuNet.tuplesToEdges — Method

Convert a path of tuples to a path of edges

QuNet.unpackStruct — Method

Unpack the attributes of a struct into attributes of a MetaDiGraph object

Properties of a MetaGraph can be anything, including structs. Normally this wouldn't be an issue but setting the weights for methods like a_star with MetaGraphs.weightfield!() requires a single weight attribute to be set. This means that the fields of a struct like "Costs" need to be individually unpacked into the MetaGraph properties

The naming convention for unpacked attributes is identical to the 'dot notation' but replaced with right arrow (→).

Example: struct Foo bar end

Foo.bar => :FooΔbar

TODO: Make sure both channels are initialised.